fix: change path for utils
