--- src/components/Utility/ProjectImage.tsx
@@ -1,19 +1,0 @@
-import marbond from "../../assets/project-1.webp";
-import suggestMeAnime from "../../assets/project-2.webp";
-import pomierskiBlog from "../../assets/project-3.webp";
-import organize from "../../assets/project-4.webp";
-import silver from "../../assets/project-5.webp";
-import myAnimeReminder from "../../assets/project-6.webp";
-import yayNay from "../../assets/project-7.webp";
-
-const projectImage = {
-  marbond,
-  suggestMeAnime,
-  pomierskiBlog,
-  organize,
-  silver,
-  myAnimeReminder,
-  yayNay,
-};
-
-export default projectImage;
